+++ NEWS.md
@@ -1,3 +1,14 @@
-# quantilogram 2.2.1
 
-This package is a preliminary version, which is devloped under R 4.1.2.
+# quantilogram 3.1.1
+
+## New features
+
+* Added new plotting functionality:
+  - `plot.crossq()`: A new method to create plots for cross-quantilogram objects.
+- `heatmap.crossq()`: A new function to generate heatmap visualizations of cross-quantilograms.
+
+## Improvements
+
+* Enhanced visualization capabilities for cross-quantilogram analysis.
+* Improved user interface for graphical representation of results.
+* Minor improvements in providing more explanation about functions in the package.